+++ src/DataSource/src/Bridge/CycleOrm/Collection/PivotedCollection.php
@@ -12,7 +12,7 @@ use spaceonfire\Collection\AbstractCollectionDecorator;
  * Attention: You should not extend this class because it will become final in the next major release
  * after the backward compatibility aliases are removed.
  *
- * @package spaceonfire\DataSource\Bridge\CycleOrm\Collection
+ * @final
  */
 class PivotedCollection extends AbstractCollectionDecorator implements PivotAwareInterface
 {