+++ src/DataSource/src/Bridge/CycleOrm/Mapper/Hydrator/Strategy/CollectionExtractorStrategy.php
@@ -17,12 +17,12 @@ use spaceonfire\DataSource\Bridge\CycleOrm\Collection\PivotAwareInterface;
 use Webmozart\Assert\Assert;
 
 /**
- * Class CollectionExtractorStrategy
+ * Class CollectionExtractorStrategy.
  *
  * Attention: You should not extend this class because it will become final in the next major release
  * after the backward compatibility aliases are removed.
  *
- * @package spaceonfire\DataSource\Bridge\CycleOrm\Mapper\Hydrator\Strategy
+ * @final
  */
 class CollectionExtractorStrategy implements StrategyInterface
 {